requirements/base.txt
Django<2.0
django-import-export==0.4.2
